server/requirements.txt
django>=4.0,<5.0
djangorestframework
psycopg2-binary
python-dotenv
django-cors-headers